third_party/blink/web_tests/fast/events/pointerevents/synthesized-middle-click-use-counter.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<body>
  <a id="link" href="#">Link</a>
  <script>
    const kSynthesizedMiddleClickAnchor = 5819;

    test(function () {
      const link = document.getElementById('link');

      internals.clearUseCounter(document, kSynthesizedMiddleClickAnchor);
      assert_false(internals.isUseCounted(document, kSynthesizedMiddleClickAnchor), "Should not be counted initially");
      const event = new MouseEvent('click', {
        button: 1 // Middle button
      });
      link.dispatchEvent(event);
      assert_true(internals.isUseCounted(document, kSynthesizedMiddleClickAnchor), "Should be counted after synthesized middle click");
    }, "Synthesized middle click on anchor should trigger UseCounter");

    test(function () {
      const link = document.getElementById('link');

      internals.clearUseCounter(document, kSynthesizedMiddleClickAnchor);
      const event = new MouseEvent('click', {
        button: 0 // Left button
      });
      link.dispatchEvent(event);
      assert_false(internals.isUseCounted(document, kSynthesizedMiddleClickAnchor), "Should not be counted after synthesized left click");
    }, "Synthesized left click on anchor should not trigger UseCounter");
  </script>
</body>
</html>
